🏗️ Put these into one file to mitigate the circular dependency warning.
